tubik: Don't install ubik.h or ubik_int.h
Both ubik and tubik try to install ubik.h and ubik_int.h. This leads
to a race during which the header file may be missing, which upsets
parallel builds. Just don't install them from tubik to solve this
issue.

This is 1.6 only - a similar fix was done as part of the libtool
work on master in 035a2ea4a1a26141508cf3fa5918d40b7be6fc90

Change-Id: I18a7dcb425b2f805656ab8ff8c06b87bcca526c7
Reviewed-on: http://gerrit.openafs.org/8808
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marc Dionne <marc.c.dionne@gmail.com>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>
Reviewed-by: Paul Smeddle <paul.smeddle@gmail.com>